Move releases table cell logic to separate component
